Simplify the data structures used for remembering APIs
Instead of having a nested object with name->version mappings we now just keep them
in a single object with 'name[/version]' keys.

Apart from simplifying the code itself it now also allows to use "group('v1')" to explicitly
get the core API (in version 'v1'), or even "group('')" to get the preferred core API version.

If a version is given both in the groupName and the versionName parameter, then the versionName
parameter wins.