Add infrastructure for providing additional options when accessing a resource (or ~ collection)
One can now chain in a call to .options({...}) before the final accessor for a resource to provide
additional options. Multiple such calls are allowed, and the last set option value wins.

The idea here is that a user could find and configure the right resource, and then hand it down
to another method for further processing but applying all options as the user wanted it.