Remove the noisy logging about unknown sub-resources
This is a "normal" situation for this client: there will always be resources
that it doesn't specifically know or understand how to make them accessible.

If that happens the user will be able to report that, and would have to provide
suitable information for how this resource appeared and/or should be used.